requirements.txt
ctransformers==0.2.5
sentence-transformers==2.6.0
pinecone-client
langchain-community
faiss-cpu
langchain==0.2.12
flask
langchainhub
protobuf==3.20.*
pypdf
langchain-pinecone==0.1.3
langchain-huggingface
pinecone-client[grpc]
python-dotenv
-e .